inference/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# PyTorch (CPU-only version)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

# Computer vision
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.24.3

# Augmentations (needed for inference_test.py imports)
albumentations==1.3.1
kornia==0.7.1

# Utilities
tqdm==4.66.1
matplotlib==3.8.2

# Logging
python-json-logger==2.0.7
